% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_modal.R
\name{mod_download_modal_server}
\alias{mod_download_modal_server}
\title{download_modal Server Functions}
\usage{
mod_download_modal_server(id, content_df, title, size)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{content_df}{reactive. A dataframe to download}

\item{title}{character. The title of the modal}

\item{size}{character. The size of the modal}
}
\description{
Module used for creating a modalDialog with a download button and options to customize the dataframe to be downloaded.
}
\examples{
library(shiny)

if (interactive()) {
  ui <- fluidPage(
    mod_download_modal_ui("download_modal_1")
  )

  server <- function(input, output, session) {
    mod_download_modal_server(
      "download_modal_1",
      content_df = reactive(iris),
      title      = "TITLE",
      size       = "s"
    )
  }

  shinyApp(ui, server)
}
}
